obligatorio ia/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openpyxl
langchain
faiss-cpu
openai